Initial refactoring to relational
