--- example.env
@@ -1,3 +1,0 @@
-MONGO_DB="tmvisdb"
-MONGO_USER='user'
-MONGO_PW='password'
