+++ src/app/sidebar.py
@@ -5,8 +5,8 @@ import logging
 
 import streamlit as st
 
-from utils.db import TaxaSelectionCriterion, Domain, Topology, DBFilter
-from utils import db
+from utils.database import TaxaSelectionCriterion, Domain, Topology, DBFilter
+from utils import database
 from utils import api
 from utils.protein_visualization import ProteinStyle, ColorScheme, VizFilter, Style
 
@@ -64,7 +64,7 @@ def create_filter_form():
             key="domain",
         )
 
-        kingdom_type = db.get_kingdom_for_domain(domain)
+        kingdom_type = database.get_kingdom_for_domain(domain)
 
         st.selectbox(
             "Select Kingdom",
@@ -102,7 +102,7 @@ def create_filter_form():
         st.number_input(
             "Select limit of shown sequences",
             1,
-            10000,
+            1000,
             value=100,
             help="As TMvisDB is a large database, you may want to set a limit for your table.",
             key="num_sequences",
